Use system gcloud by default (#12)
* use system gcloud by default

this is faster by about 20 seconds per job. this is a breaking change for any job that installs custom gcloud components, which at the moment does not appear to include any repos. this will also mean that the latest gcloud is not always installed if the default github actions runner does not have the latest version

this may be a breaking change when ubuntu/latest starts to refer to 24.04. see https://github.com/actions/runner-images/issues/7101#issuecomment-1431174447 and https://github.com/search?q=repo%3Aactions%2Frunner-images%20google-cloud&type=code

* update readme repo name